Move the sync parsing code to its own crate so it can be shared to `uniffi_parse_rs` and `uniffi_macros`
