+++ uniffi_macros/Cargo.toml
@@ -24,6 +24,7 @@ syn = { version = "2.0", features = ["full", "visit-mut"] }
 toml = ">=0.9, <2"
 uniffi_build = { path = "../uniffi_build", version = "=0.32.0", optional = true }
 uniffi_meta = { path = "../uniffi_meta", version = "=0.32.0" }
+uniffi_syn_utils = { path = "../uniffi_syn_utils", version = "=0.32.0" }
 
 [features]
 # Enable the generate_and_include_scaffolding! macro
