+++ uniffi_macros/src/fnsig.rs
@@ -132,7 +132,7 @@ impl FnSignature {
         let manual_future_output = if sig.asyncness.is_some() {
             None
         } else if let ReturnType::Type(_, ty) = &sig.output {
-            future_output_type(ty)
+            uniffi_syn_utils::future_output_type(ty)
         } else {
             None
         };
@@ -617,84 +617,6 @@ fn looks_like_result(return_type: &ReturnType) -> bool {
     false
 }
 
-/// Find a compatible future output type (`Pin<Box<dyn Future<Output = T> + Send>>`),
-/// or return `None` it none was found.
-///
-/// This matches the return type generated by `async_trait`.
-fn future_output_type(ty: &Type) -> Option<Type> {
-    let Type::Path(path) = &ty else { return None };
-
-    // First part of the path must be 'Pin'
-    let segment = &path.path.segments.last()?;
-    if segment.ident != "Pin" {
-        return None;
-    }
-
-    let syn::PathArguments::AngleBracketed(args) = &segment.arguments else {
-        return None
-    };
-
-    // Which wraps a generic argument that's another path
-    let Some(path) = args.args.iter().find_map(|arg| match arg {
-        syn::GenericArgument::Type(Type::Path(second_path)) => Some(second_path),
-        _ => None,
-    }) else {
-        return None;
-    };
-
-    // Then we should have a `Box`
-    let segment = path.path.segments.last()?;
-    if segment.ident != "Box" {
-        return None;
-    }
-
-    let syn::PathArguments::AngleBracketed(args) = &segment.arguments else {
-        return None
-    };
-
-    // And then we should have a `dyn Future` as its generic arg
-    let Some(trait_obj) = args.args.iter().find_map(|arg| match arg {
-        syn::GenericArgument::Type(Type::TraitObject(trait_obj)) => Some(trait_obj),
-        _ => None,
-    }) else {
-        return None;
-    };
-
-    // Find inner `Output` type inside `Future`
-    bounds_future_output(trait_obj.bounds.iter())
-}
-
-/// Find a `Future<Output = T>` trait bound among `bounds` and return its `Output` type.
-fn bounds_future_output<'a>(
-    bounds: impl Iterator<Item = &'a syn::TypeParamBound>,
-) -> Option<Type> {
-    for bound in bounds {
-        let syn::TypeParamBound::Trait(trait_bound) = bound else {
-            continue;
-        };
-        let Some(segment) = trait_bound.path.segments.last() else {
-            continue;
-        };
-        if segment.ident != "Future" {
-            continue;
-        }
-        if let syn::PathArguments::AngleBracketed(args) = &segment.arguments {
-            if let Some(output) = assoc_output(args) {
-                return Some(output);
-            }
-        }
-    }
-    None
-}
-
-/// Extract the `T` from an `Output = T` associated-type binding, if present.
-fn assoc_output(args: &syn::AngleBracketedGenericArguments) -> Option<Type> {
-    args.args.iter().find_map(|arg| match arg {
-        syn::GenericArgument::AssocType(assoc) if assoc.ident == "Output" => Some(assoc.ty.clone()),
-        _ => None,
-    })
-}
-
 #[derive(Debug)]
 pub(crate) enum FnKind {
     Function,
@@ -710,55 +632,4 @@ pub(crate) enum FnKind {
         self_ident: Ident,
         index: u32,
     },
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use quote::ToTokens;
-
-    /// Render `future_output_type` of `ty` back to a normalized token string, or `None`.
-    fn future_output(ty: &str) -> Option<String> {
-        let ty: Type = syn::parse_str(ty).unwrap();
-        future_output_type(&ty).map(|out| out.to_token_stream().to_string())
-    }
-
-    #[test]
-    fn plain_types_are_not_futures() {
-        assert_eq!(future_output("String"), None);
-        assert_eq!(future_output("Vec<u8>"), None);
-        // `Box` of a non-future must not be mistaken for a boxed future.
-        assert_eq!(future_output("Box<dyn Display>"), None);
-    }
-
-    #[test]
-    fn boxed_and_impl_futures_are_unwrapped() {
-        // The shape `#[async_trait]` expands `async fn ... -> String` into.
-        assert_eq!(
-            future_output("Pin<Box<dyn Future<Output = String> + Send + 'a>>"),
-            Some("String".into())
-        );
-        // Fully-qualified paths for `Pin` and `Future`.
-        assert_eq!(
-            future_output(
-                "::core::pin::Pin<Box<dyn ::core::future::Future<Output = u8> + Send>>"
-            ),
-            Some("u8".into())
-        );
-        // `Output = ()` unwraps to the unit type.
-        assert_eq!(
-            future_output("Pin<Box<dyn Future<Output = ()> + Send>>"),
-            Some("()".into())
-        );
-        // Without the `Pin` wrapper we don't consider the type to be a compatible future
-        assert_eq!(
-            future_output("Box<dyn Future<Output = Arc<Self>>>"),
-            None
-        );
-        // `impl Future` won't match either
-        assert_eq!(
-            future_output("impl Future<Output = i64>"),
-            None
-        );
-    }
 }
