+++ uniffi_parse_rs/Cargo.toml
@@ -22,6 +22,7 @@ target-lexicon = "0.13"
 thiserror = "2"
 toml = ">=0.9, <2"
 uniffi_meta = { path = "../uniffi_meta", version = "=0.32.0" }
+uniffi_syn_utils = { path = "../uniffi_syn_utils", version = "=0.32.0" }
 
 [dev-dependencies]
 expect-test = "1.4"
