+++ uniffi_parse_rs/src/functions.rs
@@ -159,7 +159,7 @@ impl ReturnType {
     pub fn parse_async(declared_async: bool, output: syn::ReturnType) -> syn::Result<(bool, Self)> {
         if !declared_async {
             if let syn::ReturnType::Type(arrow, ty) = &output {
-                if let Some(inner) = future_output_type(ty) {
+                if let Some(inner) = uniffi_syn_utils::future_output_type(ty) {
                     return Ok((
                         true,
                         Self {
@@ -206,84 +206,6 @@ impl ReturnType {
             }
         })
     }
-}
-
-/// Find a compatible future output type (`Pin<Box<dyn Future<Output = T> + Send>>`),
-/// or return `None` it none was found.
-///
-/// This matches the return type generated by `async_trait`.
-fn future_output_type(ty: &syn::Type) -> Option<syn::Type> {
-    let syn::Type::Path(path) = &ty else { return None };
-
-    // First part of the path must be 'Pin'
-    let segment = &path.path.segments.last()?;
-    if segment.ident != "Pin" {
-        return None;
-    }
-
-    let syn::PathArguments::AngleBracketed(args) = &segment.arguments else {
-        return None
-    };
-
-    // Which wraps a generic argument that's another path
-    let Some(path) = args.args.iter().find_map(|arg| match arg {
-        syn::GenericArgument::Type(syn::Type::Path(second_path)) => Some(second_path),
-        _ => None,
-    }) else {
-        return None;
-    };
-
-    // Then we should have a `Box`
-    let segment = path.path.segments.last()?;
-    if segment.ident != "Box" {
-        return None;
-    }
-
-    let syn::PathArguments::AngleBracketed(args) = &segment.arguments else {
-        return None
-    };
-
-    // And then we should have a `dyn Future` as its generic arg
-    let Some(trait_obj) = args.args.iter().find_map(|arg| match arg {
-        syn::GenericArgument::Type(syn::Type::TraitObject(trait_obj)) => Some(trait_obj),
-        _ => None,
-    }) else {
-        return None;
-    };
-
-    // Find inner `Output` type inside `Future`
-    bounds_future_output(trait_obj.bounds.iter())
-}
-
-/// Find a `Future<Output = T>` trait bound among `bounds` and return its `Output` type.
-fn bounds_future_output<'a>(
-    bounds: impl Iterator<Item = &'a syn::TypeParamBound>,
-) -> Option<syn::Type> {
-    for bound in bounds {
-        let syn::TypeParamBound::Trait(trait_bound) = bound else {
-            continue;
-        };
-        let Some(segment) = trait_bound.path.segments.last() else {
-            continue;
-        };
-        if segment.ident != "Future" {
-            continue;
-        }
-        if let syn::PathArguments::AngleBracketed(args) = &segment.arguments {
-            if let Some(output) = assoc_output(args) {
-                return Some(output);
-            }
-        }
-    }
-    None
-}
-
-/// Extract the `T` from an `Output = T` associated-type binding, if present.
-fn assoc_output(args: &syn::AngleBracketedGenericArguments) -> Option<syn::Type> {
-    args.args.iter().find_map(|arg| match arg {
-        syn::GenericArgument::AssocType(assoc) if assoc.ident == "Output" => Some(assoc.ty.clone()),
-        _ => None,
-    })
 }
 
 #[cfg(test)]
