+++ uniffi_syn_utils/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "uniffi_syn_utils"
+version = "0.32.0"
+description = "Shared syn AST helpers that are used in several crates"
+documentation = "https://mozilla.github.io/uniffi-rs"
+homepage = "https://mozilla.github.io/uniffi-rs"
+repository = "https://github.com/mozilla/uniffi-rs"
+license = "MPL-2.0"
+edition = "2021"
+keywords = ["ffi", "bindgen"]
+readme = "../README.md"
+
+[dependencies]
+syn = "2.0"
+
+[dev-dependencies]
+quote = "1"
